Copied the source code from the SourceForge repo at https://sourceforge.net/p/opendx/code/HEAD/tree/ (taken on Mar 30, 2021), which appears to be a resqued source tree from berlios.de OpenDX2 repo.
This source tree carries version tag 4.4.5, and appears to be significantly
more recent than the 4.4.4 distribution we are based on. It incorporates a
number of fixes to the SMP support under Linux, and appears to have partial
support for large arenas. At the same time, it still contains many of the
problems fixed in our trunk.

The plan is to eventually fold these patches into our trunk.